refactor, fix division by 0

+++ src/transfer.hpp
@@ -62,7 +62,7 @@ void buck::notify_transfer(const name& from, const name& to, const asset& quanti
   run(3);
 }
 
-void buck::open(const name& account, const asset& quantity, uint32_t ccr, uint32_t acr) {
+void buck::open(const name& account, const asset& quantity, uint16_t ccr, uint16_t acr) {
   check(_stat.begin() != _stat.end(), "contract is not yet initiated");
   require_auth(account);
